Ingore denounce for isready
Having a default debounce functionality is essential, but applying it to the "isready" field only slows down page loading as "isready" is called only once. 


To optimize the performance, it is wise to make the debounce feature configurable per field. However, since the model definition syntax is not yet stable (e.g., @vars vs. @handlers), for now, I suggest applying it only to the "isready" field.